Create MCE_to_Uart.asm
This code will dump the link register to UART. Useful for hunting down software triggered machine checks on the Xbox 360. Just add this as an include to your main patch file, and ensure you set the two addresses appropriately. One, the start of the machine check handler (usually around address 200 in HV), and then some free space elsewhere for the patch itself.